+++ crates/transaction-return-data/src/lib.rs
@@ -6,8 +6,8 @@ use solders_traits_core::transaction_status_boilerplate;
 
 use pyo3::prelude::*;
 use serde::{Deserialize, Serialize};
-use solana_transaction_context::TransactionReturnData as TransactionReturnDataOriginal;
-use solana_transaction_status_client_types::UiTransactionReturnData;
+use solana_transaction_context::transaction::TransactionReturnData as TransactionReturnDataOriginal;
+use solana_transaction_status_client_types::{UiReturnDataEncoding, UiTransactionReturnData};
 use solders_macros::{common_methods, richcmp_eq_only};
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize, From, Into)]
@@ -41,7 +41,11 @@ impl TransactionReturnData {
 
 impl From<TransactionReturnData> for UiTransactionReturnData {
     fn from(t: TransactionReturnData) -> Self {
-        TransactionReturnDataOriginal::from(t).into()
+        let inner = TransactionReturnDataOriginal::from(t);
+        Self {
+            program_id: inner.program_id.to_string(),
+            data: (base64::encode(inner.data), UiReturnDataEncoding::Base64),
+        }
     }
 }