+++ python/solders/__init__.py
@@ -76,4 +76,4 @@ else:
     __all__ = __all_core
 
 
-__version__ = "0.27.1"
+__version__ = "0.28.0"